Second fix for TID and NID references.

+++ src/Server.cpp
@@ -63,7 +63,7 @@ Reference getRef(const string& e2Ref) {
 string getTitanRef(const Reference& ref) {
 	ostringstream os;
 
-	unsigned long long tid = (ref.nid << 16) + ref.tid;
+	unsigned long long tid = ((unsigned long long) ref.nid << 16) + ref.tid;
 	os << ref.sid << ',' << tid;
 
 	return os.str();
